Move import arcade into test functions
conftest.py mocks pyglet.window.Window in order to test arcade.Window.
If we import arcade in any other module, the patching is not yet done
and we will be dealing with a real window, which is not what we want.

This is a temporary fix, until we can better mock arcade.Window base
class.